+++ transport/src/solana.rs
@@ -758,7 +758,7 @@ impl TransportT for SolanaTransport {
                 .into_iter()
                 .map(|p| PlayerJoin {
                     addr: p.addr.to_string(),
-                    position: p.position as usize,
+                    position: p.position,
                     balance: p.balance,
                     access_version: p.access_version,
                 })
